DABOM-222 rename: helper 디렉토리 생성 및 클래스 이동, 클래스 이름 변경
